fix profile
